fix: id_notif undefined on edit book

+++ controllers/BookReqController.js
@@ -129,14 +129,16 @@ const BookReqController = {
       };
       await client.beginTransaction();
       console.log("BEFORE", cron.getTasks());
+      let id_notif = "";
       const notif = await client.query(
         `SELECT id_notif FROM push_sched WHERE id_req = ? AND type = 'push'`,
         [id_book]
       );
-      const id_notif = notif[0][0].id_notif;
-      await client.query(`DELETE FROM push_sched WHERE id_req = ?`, [id_book]);
-      global.scheduledTasks.delete(id_notif);
-      console.log("BEFORE", cron.getTasks());
+      if (notif[0][0]) {
+        id_notif = notif[0][0].id_notif;
+        await client.query(`DELETE FROM push_sched WHERE id_req = ?`, [id_book]);
+        global.scheduledTasks.delete(id_notif);
+      }
       const [query, value] = Client.updateQuery(payload, { id_book: id_book }, "req_book");
       const updateData = await client.query(query, value);
       const q = await client.query("SELECT id_ticket from req_book where id_book = ?", [id_book]);
